Simple editor to separate component
