+++ guiv3/src/plugins/global-components.js
@@ -3,6 +3,7 @@ import Accordion from "@/components/bootstrap/Accordion";
 import AccordionItem from "@/components/bootstrap/AccordionItem";
 import SlideUpDown from "vue3-slide-up-down";
 import CodeMirror from "@/components/codemirror/CodeMirror";
+import SimpleEditor from "@/components/simpleEditor/SimpleEditor";
 
 export default function registerGlobalComponents(vueApp) {
   vueApp.component("modal", Modal);
@@ -10,4 +11,5 @@ export default function registerGlobalComponents(vueApp) {
   vueApp.component("accordion-item", AccordionItem);
   vueApp.component("slide-up-down", SlideUpDown);
   vueApp.component("codemirror", CodeMirror);
+  vueApp.component("simpleeditor", SimpleEditor);
 }
